+++ math3d/orientation.py
@@ -84,8 +84,8 @@ class Orientation(object):
             self._data = np.transpose(np.vstack((va for va in array_args)))
         elif len(args) == 0:
             self._data = np.identity(3)
-        # // Always ensure that we use float32 (single precision floats) as fundamental type
-        self._data=self._data.astype(np.float32)
+        # // Always ensure that we use float64 as fundamental type
+        self._data=self._data.astype(np.float64)
 
     def __copy__(self):
         """Copy method for creating a copy of this Orientation."""
